Additional Changes and updates pointed out by William Edor upon review
-Added usage function. User can invoke --help on terminal for usage instructions.
-Also Added a README.md file containing usage instructions.
-Checked contents of json instruction file and threw error if file is invalid  (ie if there's a missing key-value pair)
-Changed name of script to make it more descriptive
-Modularized the code, defined a main function and invoked it beneath the script
