+++ data-preprocessing/video-download-upload/ins.json
@@ -1,0 +1,9 @@
+{
+    "excel_doc_path": "NSMQ.xlsx",
+    "sheet_name": "Sheet1",
+    "video_folder_path": "videos", 
+    "video_header_name": "Links",
+    "google_drive_folder_name": "nsmq_test",
+    "google_drive_folder_ID": "1HUOn385tax6aY07XObUleCIgTd71qF-5",
+    "service_account_path": "credentials.json"
+}
